Add files via upload
Added multi output label capability.
LIMITATION: Assumes only two classes in the classifier (on or off) and decides on the output from all the estimators based on summing the 'ones' and seeing if they're greater than half the number of estimators. If so then the forest result for that output label is 'on'. This can easily be modified to handle n classes of arbitrary value by changing the voting mechanism.